==> S1/03-EFA.Rmd <==
## Exploratory factor analysis (EFA)

```{r, include=FALSE}
set.seed(2137) # set seed for replicability
```

**NOTE:** To reproduce the exploratory factor analysis for a given number of factors, modify the `factors` parameter in the header section of the parent `.Rmd` document.

### Load inputs

We load inputs required for the analysis: the preprocessed dataset, together with some helper variables (`code_to_pl`, `code_to_en`).

``` {r}
load("./02/output/dataset.RData")
```

### Subset data

We subset data for the analysis, as well as simplify the naming convention (e.g. `SOR11` instead of `ICE-168-pl.SOR11`).

```{r}
dataset = select(qdata, contains("ICE-168"))
colnames(dataset) = names(code_to_block)
```

### Convert data

We need to recode data to a format more suitable for the analysis.

``` {r}

dataset = as.data.frame(apply(dataset,2,str_trim))

dataset = dataset %>%
  mutate_all(list(~recode(.,
                           "Zdecydowanie nie" = 1,
                           "Raczej nie" = 2,
                           "Ani tak, ani nie" = 3,
                           "Raczej tak" = 4,
                           "Zdecydowanie tak" = 5)))
```

We save the resulting dataset for easier access in subsequent analysis steps:

``` {r}
save(dataset, code_to_pl, code_to_en, code_to_block, 
     file = "./03/input/dataset.RData")
```

### Correlation matrix

First, we create a correlation matrix that will be used in subsequent steps.

**NOTE:** We use `polychoric` correlations!

``` {r}
cor_matrix = POLYCHORIC_R(dataset, method = "Fox", verbose = F)
```

```{r child = if (params$factors == 7) '03-EFA-extras.Rmd'}

```

### Factor analysis

Next, we proceed to the exploratory factor analysis (EFA). 

``` {r, message = FALSE, results = 'hide'}
PA = PA_FA(cor_matrix, Nfactors=params$factors, iterpaf=250, rotate="PROMAX", Ncases=nrow(dataset), ppower = 4, verbose=F)
```

We extract relevant information for later use:

``` {r, message = FALSE, results = 'hide'}
phi = PA$phi
rotated = PA$pattern
totvarexplNOROT1 = PA$totvarexplNOROT1
totvarexplROT = as.data.frame(PA$totvarexplROT)
```

We inspect eigenvalues, proportion of variance explained by each factor, as well as cumulative proportion of variance:

``` {r}
kable(head(totvarexplNOROT1, params$factors))
```

And also total variance explained after the rotation:

``` {r}
colnames(totvarexplROT) = c("Total Variance Expl. After Rotation")
kable(totvarexplROT)
```

### Thresholding {.tabset}

Based on the factor loadings, we need to decide which items should be included to the final version of the questionnaire.

We clean the output based on two criteria (both criteria must be met):

* `criterion1`: for any item to be included to a given factor, its corresponding factor loading must be greater than a certain threshold value
* `criterion2`: item's factor loading for this factor must be sufficiently higher than its factor loadings for the remaining factors.

This step is done to facilitate the decision process.

We can manipulate `criterion1` and `criterion2` in the code below, to see how many items end up loading each factor at different threshold values.

``` {r}
thresholded = matrix(, nrow = nrow(rotated), ncol = ncol(rotated))
colnames(thresholded) = colnames(rotated)
rownames(thresholded) = rownames(rotated)

best = apply(abs(rotated), 1, order, decreasing=TRUE)

for (i in 1:nrow(rotated)) {
  
  tmp = unname(rotated[i,])
  idx = unname(best[,i])
  
  max1 = idx[1]
  max2 = idx[2]
  
  criterion1 = abs(tmp[max1]) > 0.5
  criterion2 = ( abs(tmp[max1]) - abs(tmp[max2]) ) > 0.3
  
  test = rep(FALSE, length(tmp))
  test[max1] = criterion1 & criterion2
  
  thresholded[i,] = ifelse(test, tmp, NA)
}
```

Assuming the above criteria, the resulting number of items in each factor is as follows:

```{r}
nitems = matrix(nrow = 1, ncol = ncol(thresholded))
colnames(nitems) = colnames(thresholded)
rownames(nitems) = "Number of items"

for (i in 1:ncol(thresholded)) {
  nitems[1,i] = sum(!is.na(thresholded[,i]))
}

kable(nitems)
```

The factors correlations are as follows:

```{r}
kable(phi, digits = 2)
```

The table below displays the factor loadings that passed the above criteria. Use the tabs below to switch between thresholded and unthresholded views.

#### Thresholded

``` {r}
kable(thresholded, digits = 2)
```

#### Unthresholded

``` {r}
kable(rotated, digits = 2)
```

### Lists of best loading items in Polish {.tabset}

Items (item codes + item texts in Polish) that passed the above criteria are listed below. Use the tabs to view items for different factors.

``` {r, echo=FALSE}
factor_pl = NULL

for (i in 1:ncol(thresholded)) {
  idx_of_loading_items = !is.na(thresholded[,i])
  codes_of_loading_items = rownames(thresholded)[idx_of_loading_items]
  out = code_to_pl[codes_of_loading_items]
  out = as.data.frame(out)
  colnames(out) = "Best items"

  factor_pl[i] = list(out)
}
```

``` {r, results='asis', echo=FALSE}
for (i in 1:length(factor_pl)) {

  cat("\n")
  cat(sprintf("#### Factor %s \n", i))
  cat("\n")

  print(kable(factor_pl[[i]]))

  cat("\n")
}
```

### Lists of best loading items in English {.tabset}

Items (item codes + item texts in English) that passed the above criteria are listed below. Use the tabs to view items for different factors.

``` {r, echo=FALSE}
factor_en = NULL

for (i in 1:ncol(thresholded)) {
  idx_of_loading_items = !is.na(thresholded[,i])
  codes_of_loading_items = rownames(thresholded)[idx_of_loading_items]
  out = code_to_en[codes_of_loading_items]
  out = as.data.frame(out)
  colnames(out) = "Best items"

  factor_en[i] = list(out)
}
```

``` {r, results='asis', echo=FALSE}
for (i in 1:length(factor_en)) {

  cat("\n")
  cat(sprintf("#### Factor %s \n", i))
  cat("\n")

  print(kable(factor_en[[i]]))

  cat("\n")
}
```

### Prepare outputs

``` {r}
output = data.frame(matrix(ncol = 5, nrow = 0))

for (i in 1:length(factor_pl)) {
  fcodes = rownames(factor_pl[[i]])
  
  col1 = as.character(rep(i,length(fcodes)))
  col2 = as.character(fcodes)
  col3 = as.character(code_to_pl[fcodes])
  col4 = as.character(code_to_en[fcodes])
  col5 = as.character(lapply(thresholded[fcodes,i], round, 2))

  out = as.data.frame(cbind(col1, col2, col3, col4, col5))
  output = rbind(output,out)
}

colnames(output) = c("FACTOR", "CODE", "ITEM TEXT", "ITEM TEXT EN",	"LOAD")
```

### Save outputs

Save outputs required for further step(s).

``` {r}
fname = paste("./03/output/items-per-factor-", as.character(params$factors),".tsv", sep="")
write.table(output, file=fname, row.names = F, sep='\t', fileEncoding = "UTF-8")
```
